Add an analysis + plot to an experiment in Pluto

Uploads an HTML or image plot with optional information to a Pluto experiment. Optionally upload tabular results as well.

Usage

pluto_add_experiment_plot(
  experiment_id,
  display_file_path,
  results_file_path = NULL,
  script_file_path = NULL,
  analysis_name = NULL,
  analysis_methods = NULL
)

Arguments

experiment_id

Pluto experiment ID

display_file_path

Path to the .html or image file containing a plot to push to Pluto

results_file_path

Optional, path to the .csv file containing results to push to Pluto

script_file_path

Optional, path to the script file used to generate the plot

analysis_name

String, a title for the new plot

analysis_methods

String, a description of the methods used for making the plot